move totalDonation to onetime
